+++ src/laboratory/static/laboratory/css/laboratory.css
@@ -57,3 +57,7 @@ ul, #ullabroom {
     font-weight: bold;
 }
 
+#detail_modal_container{
+    margin: 0;
+}
+
